integrated - working visually, not fully tested
