+++ .vscode/settings.json
@@ -7,6 +7,6 @@
         "/Users/noahpearsonkramer/Desktop/mysql-connector-java-8.0.28/**/*.jar",
         "c:\\Program Files\\MySQL\\Connector J 8.0\\mysql-connector-java-8.0.28.jar"
     ],
-    "java.jdt.ls.vmargs": "-XX:+UseParallelGC -XX:GCTimeRatio=4 -XX:AdaptiveSizePolicyWeight=90 -Dsun.zip.disableMemoryMapping=true -Xmx4G -Xms100m",
+    "java.jdt.ls.vmargs": "-XX:+UseParallelGC -XX:GCTimeRatio=4 -XX:AdaptiveSizePolicyWeight=90 -Dsun.zip.disableMemoryMapping=true -Xmx8G -Xms100m",
     "java.dependency.packagePresentation": "hierarchical"
 }
