+++ Admin/InventoryManagerGUI.java
@@ -1,3 +1,4 @@
+package Admin;
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.event.*;
